changed hourly rate field to decimal type on Tutors Table

+++ Models/Tutor.cs
@@ -6,7 +6,7 @@ namespace tutorfy_backend.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int HourlyRate { get; set; }
+        public decimal HourlyRate { get; set; }
         public string ZipCode { get; set; }
         public bool IsActivated { get; set; }
         public bool IsProfileCompleted { get; set; }
